+++ packages/table/src/utils/cellRenderToFromItem.tsx
@@ -1,6 +1,6 @@
-import React, { memo, useCallback, useEffect, useMemo, useState } from 'react';
+import React, { memo, useCallback, useContext, useEffect, useMemo, useState } from 'react';
 import type { ProFormFieldProps } from '@ant-design/pro-form';
-import { ProFormField, ProFormDependency } from '@ant-design/pro-form';
+import { FieldContext, ProFormField, ProFormDependency } from '@ant-design/pro-form';
 import type { ProFieldEmptyText } from '@ant-design/pro-field';
 import type { ProFieldValueType, ProSchemaComponentTypes } from '@ant-design/pro-utils';
 import { isDeepEqualReact } from '@ant-design/pro-utils';
@@ -45,16 +45,28 @@ type CellRenderFromItemProps<T> = {
   prefixName?: string;
   counter: ReturnType<typeof useContainer>;
   proFieldProps: ProFormFieldProps;
+  subName: string[];
 };
 
 const CellRenderFromItem = <T,>(props: CellRenderFromItemProps<T>) => {
+  const formContext = useContext(FieldContext);
   /**
    * memo cannot use generics type, so wrap it
    */
   const Component = useMemo(
     () =>
       memo<CellRenderFromItemProps<T>>(
-        ({ columnProps, prefixName, text, counter, rowData, index, recordKey, proFieldProps }) => {
+        ({
+          columnProps,
+          prefixName,
+          text,
+          counter,
+          rowData,
+          index,
+          recordKey,
+          subName,
+          proFieldProps,
+        }) => {
           const { editableForm } = counter;
           const key = recordKey || index;
           const [name, setName] = useState<React.Key[]>([]);
@@ -66,6 +78,7 @@ const CellRenderFromItem = <T,>(props: CellRenderFromItemProps<T>) => {
           useEffect(() => {
             const value = spellNamePath(
               prefixName,
+              subName,
               prefixName ? index : key,
               columnProps?.key ?? columnProps?.dataIndex ?? index,
             );
@@ -89,10 +102,21 @@ const CellRenderFromItem = <T,>(props: CellRenderFromItemProps<T>) => {
 
           const InlineItem = useCallback<React.FC<any>>(
             ({ children, ...restProps }) => (
-              <InlineErrorFormItem key={key} errorType="popover" name={name} {...restProps}>
+              <InlineErrorFormItem
+                popoverProps={{
+                  getPopupContainer:
+                    formContext.getPopupContainer ||
+                    (() => counter.rootDomRef.current || document.body),
+                }}
+                key={key}
+                errorType="popover"
+                name={name}
+                {...restProps}
+              >
                 {children}
               </InlineErrorFormItem>
             ),
+            // eslint-disable-next-line react-hooks/exhaustive-deps
             [key, name],
           );
 
@@ -224,7 +248,9 @@ function cellRenderToFromItem<T>(config: CellRenderFromItemProps<T>): React.Reac
 
   const columnKey = columnProps?.key || columnProps?.dataIndex?.toString();
 
-  /** 生成公用的 proField dom 配置 */
+  /**
+   * 生成公用的 proField dom 配置
+   */
   const proFieldProps: ProFormFieldProps = {
     valueEnum: runFunction<[T | undefined]>(columnProps?.valueEnum, rowData),
     request: columnProps?.request,