+++ packages/table/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.68.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.68.2...@ant-design/pro-table@2.68.3) (2022-03-22)
+
+**Note:** Version bump only for package @ant-design/pro-table
+
+## [2.68.2](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.68.1...@ant-design/pro-table@2.68.2) (2022-03-18)
+
+### Bug Fixes
+
+- **table:** fix inlint error render dom ([bfeb81a](https://github.com/ant-design/pro-components/commit/bfeb81afc6bc4305b3d7ef2b84c0afc089dad7a0))
+- **table:** support ghost ([7bec0df](https://github.com/ant-design/pro-components/commit/7bec0dfe373884ec18c10e965c1a9f2c1fff9605))
+
+## [2.68.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.68.0...@ant-design/pro-table@2.68.1) (2022-03-16)
+
+**Note:** Version bump only for package @ant-design/pro-table
+
 # [2.68.0](https://github.com/ant-design/pro-components/compare/@ant-design/pro-table@2.67.1...@ant-design/pro-table@2.68.0) (2022-03-16)
 
 ### Features